Add MinGW64/MSYS2 cross-platform build support
Port the macOS-only build to compile on Windows with MinGW64:
- Add CMakeLists.txt for cmake-based builds with StellarSolver static linking
- Fix Qt6 API usage (remove setCodec, use errorOccurred signal)
- Fix std::min type mismatch (int vs qsizetype) on 64-bit MinGW
- Platform-specific monospace fonts and astrometry index paths
- Remove hardcoded macOS paths from main.cpp
- Make .pro file cross-platform (macOS, Windows, Linux)
- Add stellarsolver_static.h to suppress dllimport for static linking
- Document full MinGW64 build process including wcslib/stellarsolver

+++ AlignmentController.cpp
@@ -590,7 +590,7 @@ void AlignmentController::onPlateSolveCompleted(const PlateSolveResult &result)
             setState(Error);
             return;
         }
-        m_numStarsRequested = std::min(m_numStarsRequested, m_visibleStars.size());
+        m_numStarsRequested = std::min(m_numStarsRequested, static_cast<int>(m_visibleStars.size()));
 
         setState(SelectingStar);
         selectNextStar();
